+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changes
 =======
 
+0.1.3
+-----
+- `fragmentmapbed` routine that converts a fragment map to the BED 
+format;
+- `fragmentmapstat` routine that shows summary on a fragment map.
+
 0.1.2
 -----
 - `simulator` routine that simulates random genome fragments and its 
@@ -8,6 +14,6 @@ chromosomes.
 
 0.1.1
 -----
-- `assemble`, `fragmentmap` and `transfer` routines implemented 
-- logging messages for `fragmentmap` and `transfer`
+- `assemble`, `fragmentmap` and `transfer` routines implemented; 
+- logging messages for `fragmentmap` and `transfer`.